Add build action

+++ build.py
@@ -1,14 +1,15 @@
 toCopy = (
     "index.html",
     "nav_bg.png",
-    "uni_sans.otf",
-    "calibril.ttf"
+#    "uni_sans.otf",
+#    "calibril.ttf"
 )
 
 import shutil
 import os
 
-shutil.rmtree("out")
+if os.path.exists("out"):
+    shutil.rmtree("out")
 os.makedirs("out")
